// lib/types.ts
export type Frontmatter = {
  proficiency_bonus?: number;
  level?: number;
  [key: string]: any; // Allow other frontmatter properties
};

export type TTRPGBlock = {
  type: 'dnd' | 'daggerheart';
}

export type DNDAbilityBlock = {
  type: 'dnd';
  abilities: DNDAbilityScores;
  bonuses: GenericBonus[];
  proficiencies: string[];
}

export type DaggerHeartAbilityBlock = {
  type: 'daggerheart';
  abilities: DaggerHeartAbilityScores;
}

export type AbilityBlock = DNDAbilityBlock | DaggerHeartAbilityBlock;

export type DNDAbilityScores = {
  strength: number;
  dexterity: number;
  constitution: number;
  intelligence: number;
  wisdom: number;
  charisma: number;
};

export type DaggerHeartAbilityScores = {
  agility: number;
  strength: number;
  finesse: number;
  instinct: number;
  presence: number;
  knowledge: number;
};

// An GenericBonus is an additional property for the ability block
// that allows for custom additions to score points. This helps users
// add things from feats, or other sources that might modify the score.
export type GenericBonus = {
  name: string;
  target: keyof DNDAbilityScores | keyof DaggerHeartAbilityScores;
  value: number;
  modifies?: "saving_throw" | "score"; // Defaults to 'saving_throw'
};

export type StatItem = {
  label: string;
  value: string | number;
  sublabel?: string;
};

export type StatsBlock = {
  items: StatItem[];
  grid?: {
    columns?: number;
  };
  dense?: boolean;
};

export type SkillsBlock = {
  proficiencies: string[];
  expertise: string[];
  half_proficiencies: string[];
  bonuses: SkillsBlockBonus[];
};

export type SkillsBlockBonus = GenericBonus;